% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForwardKinematics.R
\name{.ur3_forward_kinematics}
\alias{.ur3_forward_kinematics}
\title{Perform and plot iterations of forward kinematics}
\usage{
.ur3_forward_kinematics(joint_angles)
}
\arguments{
\item{joint_values}{}
}
\value{
A list containing the end-effector position and orientation  position -
1x3 end_effector_position vector, orientation = 1x3 end_effector_orientation vector
}
\description{
Perform and plot iterations of forward kinematics
}
\examples{
joint_values <- c(0, 0, 0, 0, 0, 0)
ur3_forward_kinematics(joint_values)
}
