% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_topic_info.R
\name{extract_topic_info}
\alias{extract_topic_info}
\title{Extract the topic information from the bag file and parse all the topics through the format_msg_csv function, generating a list of usable data frames}
\usage{
extract_topic_info(bag_name = NULL)
}
\arguments{
\item{bag_name}{\itemize{
\item The name of the bag file recorded by ROS2, defaults to ur3_move, which is included in the vignettes folder
}}
}
\value{
A list with data frames of recorded topic information
\item{output_data_frame\link{joint_states}}
\item{joint_states topic}{Position, velocity, and effort values for 6-DOF robotic manipulator.}
\item{output_data_frame\link{ur_controller_state}}{Error, desired, and actual position values for 6-DOF robotic manipulator.}
\item{ur_controller_state topic}{Error, desired, and actual position values for 6-DOF robotic manipulator.}
}
\description{
Extract the topic information from the bag file and parse all the topics through the format_msg_csv function, generating a list of usable data frames
}
\examples{
extract_topic_info("ur3_move")
extract_topic_info("ur3_move_two")


}
