% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_topic_info.R, R/format_msg_csv.R
\name{format_msg_csv}
\alias{format_msg_csv}
\title{Format the echoed message data to a data frame to be used for plotting}
\usage{
format_msg_csv(bag_name = NULL, topic_name = NULL)

format_msg_csv(bag_name = NULL, topic_name = NULL)
}
\arguments{
\item{bag_name}{\itemize{
\item The name of the bag file recorded by ROS2, defaults to ur3_move, which is included in the vignettes folder
}}

\item{topic_name}{\itemize{
\item The name of the topic that was recorded, defaults to joint_states. Messages used for the topics are included in the src folder
}}
}
\value{
joint_data_frame - Data frame for a single topic ( 2 available topics )
\item{\link{joint_states}} - {Position, velocity, and effort values for 6-DOF robotic manipulator.}
\item{\link{ur_controller_state}} - {Error, desired, and actual position values for 6-DOF robotic manipulator.}

joint_data_frame - Data frame for a single topic ( 2 available topics )
\item{\link{joint_states}} - {Position, velocity, and effort values for 6-DOF robotic manipulator.}
\item{\link{ur_controller_state}} - {Error, desired, and actual position values for 6-DOF robotic manipulator.}
}
\description{
Format the echoed message data to a data frame to be used for plotting

Format the echoed message data to a data frame to be used for plotting
}
\examples{
data_frame_JS <- format_msg_csv("ur3_move", "joint_states")
data_frame_JTCS <- format_msg_csv("ur3_move_two", "ur_controller_state")
data_frame_JS <- format_msg_csv("ur3_move", "joint_states")
data_frame_JTCS <- format_msg_csv("ur3_move_two", "ur_controller_state")
}
