% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForwardKinematics.R
\name{update_plot}
\alias{update_plot}
\title{Forward Kinematics for UR3 Robot Arm}
\usage{
update_plot(selected_data, time_interval = 500)
}
\arguments{
\item{selected_data}{\itemize{
\item The data frame of joint values returned from a topic
}}

\item{time_interval}{}
}
\value{
A 3D plot of the end effector trajectory
}
\description{
Forward Kinematics for UR3 Robot Arm
}
\examples{
selected_data <- format_msg_csv("ur3_move", "joint_states")
update_plot(selected_data)

}
